+++ Network/Telegram/API/Bot/Object/Chat.hs
@@ -9,23 +9,12 @@ import "base" Data.Functor ((<$>))
 import "base" Text.Show (Show)
 import "text" Data.Text (Text)
 
-import Network.Telegram.API.Bot.Property.Identifiable
-	(Identifiable (identificator), Identificator)
-
 data Chat
 	= Private Int64
 	| Group Int64 Text
 	| Supergroup Int64 Text
 	| Channel Int64 Text
 	deriving Show
-
-type instance Identificator Chat = Int64
-
-instance Identifiable Chat where
-	identificator (Private i) = i
-	identificator (Group i _) = i
-	identificator (Supergroup i _) = i
-	identificator (Channel i _) = i
 
 instance FromJSON Chat where
 	parseJSON = withObject "Chat" $ \v -> v .: "type" >>= \case