+++ CHANGELOG.md
@@ -29,3 +29,4 @@
 
 # 0.1.5
 * Remove `Internal` module, move its content to `Endpoint`
+* Remove `Drop` phantom datatype, replace it with `Post`